refactor: move to config/shared
