+++ config/shared/weapons.lua
