+++ shared/main.lua
@@ -1,8 +1,8 @@
 local qbShared = {}
 qbShared.ForceJobDefaultDutyAtLogin = true -- true: Force duty state to jobdefaultDuty | false: set duty state from database last saved
-qbShared.Locations = require 'data.locations'
-qbShared.Vehicles = require 'data.vehicles'
-qbShared.Weapons = require 'data.weapons'
+qbShared.Locations = require 'config.shared.locations'
+qbShared.Vehicles = require 'config.shared.vehicles'
+qbShared.Weapons = require 'config.shared.weapons'
 
 ---@type table<number, Vehicle>
 qbShared.VehicleHashes = {}
